[DPE-7941] Add GPU support (#125)
* feat: Add configuration options needed for gpu support

* fix: Fix k8s_node_selectors type issue

* feat: Set engine timeout to 1M in testing profile

* feat: Add GPU support

* fix: Do not remove test bucket if --keep-models is set

* tests: Initial addition of GPU test

* fix: Fix gpu capacity access

* fix: Fix gpu pod template path

* tests: Update charm revisions to latest stables

* ci: Add initial gpu integration test ci on aws

* tests: Remove obsolete add-config action calls to hub

* tests: Fix ssc base

* ci: Use --classic flag to install non-strict microk8s

* ci: Fix ec2 setup

* refactor: Rename gpu-enable config option

* feat: Add gpu-pinned-memory config option

* chore: Update OCI resources

* refactor: Remove ttlSecondsAfterFinished from pod template

* refactor: Change gpu-engine-executors-limit use-all to -1

* feat: Block charm if more GPU requested than capacity

* chore: Update OCI resources and remove pod template write

* tests: Assert that the driver pod does not request GPU

+++ tests/integration/conftest.py
@@ -58,15 +58,15 @@ def charm_versions() -> IntegrationTestsCharms:
     return IntegrationTestsCharms(
         s3=TestCharm(
             name="s3-integrator",
-            channel="edge",
-            revision=41,
+            channel="1/stable",
+            revision=145,
             base="ubuntu@22.04",
             alias="s3",
         ),
         metastore_db=TestCharm(
             name="postgresql-k8s",
             channel="14/stable",
-            revision=281,
+            revision=495,
             base="ubuntu@22.04",
             alias="metastore",
             trust=True,
@@ -74,15 +74,15 @@ def charm_versions() -> IntegrationTestsCharms:
         auth_db=TestCharm(
             name="postgresql-k8s",
             channel="14/stable",
-            revision=281,
+            revision=495,
             base="ubuntu@22.04",
             alias="auth-db",
             trust=True,
         ),
         integration_hub=TestCharm(
             name="spark-integration-hub-k8s",
-            channel="latest/edge",
-            revision=43,
+            channel="3/stable",
+            revision=67,
             base="ubuntu@22.04",
             alias="integration-hub",
             trust=True,
@@ -96,9 +96,9 @@ def charm_versions() -> IntegrationTestsCharms:
         ),
         tls=TestCharm(
             name="self-signed-certificates",
-            channel="latest/stable",
-            revision=163,  # FIXME (certs): Unpin the revision once the charm is fixed
-            base="ubuntu@22.04",
+            channel="1/stable",
+            revision=317,
+            base="ubuntu@24.04",
             alias="self-signed-certificates",
         ),
         data_integrator=TestCharm(
@@ -112,7 +112,8 @@ def charm_versions() -> IntegrationTestsCharms:
 
 
 @pytest.fixture(scope="module")
-def s3_bucket_and_creds():
+def s3_bucket_and_creds(request: pytest.FixtureRequest):
+    keep_models = bool(request.config.getoption("--keep-models"))
     logger.info("Fetching S3 credentials from minio.....")
 
     fetch_s3_output = (
@@ -163,12 +164,13 @@ def s3_bucket_and_creds():
         "path": TEST_PATH_NAME,
     }
 
-    logger.info("Tearing down test bucket...")
-    for obj in test_bucket.objects.all():
-        # We need to iterate over keys because delete_objects (plural) has mandatory checksum
-        obj.delete()
+    if not keep_models:
+        logger.info("Tearing down test bucket...")
+        for obj in test_bucket.objects.all():
+            # We need to iterate over keys because delete_objects (plural) has mandatory checksum
+            obj.delete()
 
-    test_bucket.delete()
+        test_bucket.delete()
 
 
 @pytest.fixture(scope="module")
